NEW: CTC clock trim driver, seperated peripheral pin configuration from the general peripheral config. Various improvements to ADC and FMC code. Readability and API consistancy updates. Remove the need to duplicate read/write register functions by handling them similarly to read/write bit(s). This removes a lot of duplicate code. Add some new hekper functions and other general code cleanup. FIXED: GPIO mode and control bits being swapped. Fixed cases where a preipheral may have been accessed before clock initialization.

+++ Source/DMA/dma_config.hpp
@@ -20,7 +20,7 @@ enum class DMA_Base {
     DMA1_BASE,
 };
 
-static constexpr unsigned int DMA_baseAddress[] = {
+static constexpr uintptr_t DMA_baseAddress[] = {
     0x40020000,
     0x40020400,
 };
